requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.1
defusedxml==0.7.1
dj-rest-auth==2.1.11
Django==4.0
django-allauth==0.52.0
django-cors-headers==3.10.0
djangorestframework==3.13.0
idna==3.4
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
